refactor: change float type to double for more precision

+++ src/main/java/Level1/Product.java
@@ -2,9 +2,9 @@ package Level1;
 
 public class Product {
     private String name;
-    private float price;
+    private double price;
 
-    public Product(String name, float price) {
+    public Product(String name, double price) {
         this.name = name;
         this.price = price;
     }
@@ -13,7 +13,7 @@ public class Product {
         return name;
     }
 
-    public float getPrice() {
+    public double getPrice() {
         return price;
     }
 }
